+++ go-controller/pkg/ovn/network_controller_manager.go
@@ -1,0 +1,180 @@
+package ovn
+
+import (
+	"fmt"
+	"sync"
+
+	"github.com/containernetworking/cni/pkg/types"
+	libovsdbclient "github.com/ovn-org/libovsdb/client"
+	ovncnitypes "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/cni/types"
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/factory"
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/kube"
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/libovsdbops"
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/metrics"
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/nad-controller"
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/nbdb"
+	ovntypes "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/types"
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/util"
+
+	clientset "k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/tools/record"
+	"k8s.io/klog/v2"
+)
+
+// networkControllerManager structure is the object manages all controllers for all networks
+// Implements nad_controller.NetworkControllerManager interface
+type networkControllerManager struct {
+	client       clientset.Interface
+	kube         *kube.KubeOVN
+	watchFactory *factory.WatchFactory
+	podRecorder  *metrics.PodRecorder
+	// event recorder used to post events to k8s
+	recorder record.EventRecorder
+	// libovsdb northbound client interface
+	nbClient libovsdbclient.Client
+	// libovsdb southbound client interface
+	sbClient libovsdbclient.Client
+	// has SCTP support
+	SCTPSupport bool
+	// Supports multicast?
+	multicastSupport bool
+
+	stopChan chan struct{}
+	wg       *sync.WaitGroup
+}
+
+func (cm *networkControllerManager) NewNetworkController(nInfo util.NetInfo,
+	netConfInfo util.NetConfInfo) (nad_controller.NetworkController, error) {
+	cnci := cm.newCommonNetworkControllerInfo()
+	topoType := netConfInfo.TopologyType()
+	switch topoType {
+	case ovntypes.Layer3Topology:
+		return NewSecondaryLayer3NetworkController(cnci, nInfo, netConfInfo), nil
+	case ovntypes.Layer2Topology:
+		return NewSecondaryLayer2NetworkController(cnci, nInfo, netConfInfo), nil
+	}
+	return nil, fmt.Errorf("topology type %s not supported", topoType)
+}
+
+// newDummyNetworkController creates a dummy network controller used to clean up specific network
+func (cm *networkControllerManager) newDummyNetworkController(topoType, netName string) (nad_controller.NetworkController, error) {
+	cnci := cm.newCommonNetworkControllerInfo()
+	netInfo := util.NewNetInfo(&ovncnitypes.NetConf{NetConf: types.NetConf{Name: netName}, Topology: topoType})
+	switch topoType {
+	case ovntypes.Layer3Topology:
+		return NewSecondaryLayer3NetworkController(cnci, netInfo, &util.Layer3NetConfInfo{}), nil
+	case ovntypes.Layer2Topology:
+		return NewSecondaryLayer2NetworkController(cnci, netInfo, &util.Layer2NetConfInfo{}), nil
+	}
+	return nil, fmt.Errorf("topology type %s not supported", topoType)
+}
+
+// Find all the OVN logical switches/routers for the secondary networks
+func findAllSecondaryNetworkLogicalEntities(nbClient libovsdbclient.Client) ([]*nbdb.LogicalSwitch,
+	[]*nbdb.LogicalRouter, error) {
+	p1 := func(item *nbdb.LogicalSwitch) bool {
+		_, ok := item.ExternalIDs[ovntypes.NetworkExternalID]
+		return ok
+	}
+	nodeSwitches, err := libovsdbops.FindLogicalSwitchesWithPredicate(nbClient, p1)
+	if err != nil {
+		klog.Errorf("Failed to get all logical switches of secondary network error: %v", err)
+		return nil, nil, err
+	}
+	p2 := func(item *nbdb.LogicalRouter) bool {
+		_, ok := item.ExternalIDs[ovntypes.NetworkExternalID]
+		return ok
+	}
+	clusterRouters, err := libovsdbops.FindLogicalRoutersWithPredicate(nbClient, p2)
+	if err != nil {
+		klog.Errorf("Failed to get all distributed logical routers: %v", err)
+		return nil, nil, err
+	}
+	return nodeSwitches, clusterRouters, nil
+}
+
+func (cm *networkControllerManager) CleanupDeletedNetworks(allControllers []nad_controller.NetworkController) error {
+	existingNetworksMap := map[string]struct{}{}
+	for _, oc := range allControllers {
+		existingNetworksMap[oc.GetNetworkName()] = struct{}{}
+	}
+
+	// Get all the existing secondary networks and its logical entities
+	switches, routers, err := findAllSecondaryNetworkLogicalEntities(cm.nbClient)
+	if err != nil {
+		return err
+	}
+
+	staleNetworkControllers := map[string]nad_controller.NetworkController{}
+	for _, ls := range switches {
+		netName := ls.ExternalIDs[ovntypes.NetworkExternalID]
+		if _, ok := existingNetworksMap[netName]; ok {
+			// network still exists, no cleanup to do
+			continue
+		}
+		// TopologyExternalID always co-exists with NetworkExternalID
+		topoType := ls.ExternalIDs[ovntypes.TopologyExternalID]
+		// Create dummy network controllers to clean up logical entities
+		klog.V(5).Infof("Found stale %s network %s", topoType, netName)
+		if oc, err := cm.newDummyNetworkController(topoType, netName); err == nil {
+			staleNetworkControllers[netName] = oc
+			continue
+		}
+	}
+	for _, lr := range routers {
+		netName := lr.ExternalIDs[ovntypes.NetworkExternalID]
+		if _, ok := existingNetworksMap[netName]; ok {
+			// network still exists, no cleanup to do
+			continue
+		}
+		// TopologyExternalID always co-exists with NetworkExternalID
+		topoType := lr.ExternalIDs[ovntypes.TopologyExternalID]
+		// Create dummy network controllers to clean up logical entities
+		klog.V(5).Infof("Found stale %s network %s", topoType, netName)
+		if oc, err := cm.newDummyNetworkController(topoType, netName); err == nil {
+			staleNetworkControllers[netName] = oc
+			continue
+		}
+	}
+
+	for netName, oc := range staleNetworkControllers {
+		klog.Infof("Cleanup entities for stale network %s", netName)
+		err = oc.Cleanup(netName)
+		if err != nil {
+			klog.Errorf("Failed to delete stale OVN logical entities for network %s: %v", netName, err)
+		}
+	}
+	return nil
+}
+
+// NewNetworkControllerManager creates a new OVN controller manager to manage all the controller for all networks
+func NewNetworkControllerManager(ovnClient *util.OVNClientset, identity string, wf *factory.WatchFactory,
+	libovsdbOvnNBClient libovsdbclient.Client, libovsdbOvnSBClient libovsdbclient.Client,
+	recorder record.EventRecorder, wg *sync.WaitGroup) *networkControllerManager {
+	podRecorder := metrics.NewPodRecorder()
+
+	cm := &networkControllerManager{
+		client: ovnClient.KubeClient,
+		kube: &kube.KubeOVN{
+			Kube:                 kube.Kube{KClient: ovnClient.KubeClient},
+			EIPClient:            ovnClient.EgressIPClient,
+			EgressFirewallClient: ovnClient.EgressFirewallClient,
+			CloudNetworkClient:   ovnClient.CloudNetworkClient,
+		},
+		stopChan:     make(chan struct{}),
+		watchFactory: wf,
+		recorder:     recorder,
+		nbClient:     libovsdbOvnNBClient,
+		sbClient:     libovsdbOvnSBClient,
+		podRecorder:  &podRecorder,
+
+		wg: wg,
+	}
+	return cm
+}
+
+// newBaseNetworkController creates and returns the base controller
+func (cm *networkControllerManager) newCommonNetworkControllerInfo() *CommonNetworkControllerInfo {
+	return NewCommonNetworkControllerInfo(cm.client, cm.kube, cm.watchFactory, cm.recorder, cm.nbClient,
+		cm.sbClient, cm.podRecorder, cm.SCTPSupport, cm.multicastSupport)
+}
